Subtraction rule cannot trigger when a division is present
